NH-43527: add support of activerecord 7.1 for non-rails app
