+++ lib/solarwinds_apm/support.rb
@@ -22,14 +22,11 @@ if SolarWindsAPM::Config[:tag_sql]
     if ::Rails.version < '7'
       require_relative './support/swomarginalia/railtie'
     elsif ::Rails.version >= '7'
-      # User has to define in their config/environments:
-      # config.active_record.query_log_tags = [ 
-      #   { 
-      #     tracecontext: -> { 
-      #       SolarWindsAPM::SWOMarginalia::Comment.traceparent 
-      #     } 
-      #   }
-      # ]
+      # User has to configure application.rb with
+      # config.active_record.query_log_tags_enabled = true
+      #
+      # User has to configure environments/*.rb with
+      # config.active_record.query_log_tags = [{ traceparent: -> { SolarWindsAPM::SWOMarginalia::Comment.traceparent } }]
       SolarWindsAPM.logger.info {"In Rails 7, tag tracecontext on a query by including SolarWindsAPM::SWOMarginalia::Comment.traceparent as function in config.active_record.query_log_tags."}
       SolarWindsAPM.logger.info {"For more information, please check https://api.rubyonrails.org/classes/ActiveRecord/QueryLogs.html"}
       require_relative './support/swomarginalia/comment'
